chore: Update Main class
Description:
This commit updates the `Main` class to incorporate the BusDAOImpl for retrieving bus entities from the database. The changes include importing the BusDAOImpl class, obtaining all buses from the database, and passing the list of buses to the FloydWarshall constructor.

Changes:
- Imported the BusDAOImpl class.
- Added a new BusDAOImpl instance to fetch bus entities.
- Retrieved all buses from the database using the busDAO.getEntities() method.
- Passed the list of buses to the FloydWarshall constructor.

Context:
The modification ensures that the Main class is up-to-date with the project requirements, taking into account the retrieval of buses for the Floyd-Warshall algorithm.